+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "x6-html-shape",
-  "version": "0.3.6",
+  "version": "0.3.7",
   "type": "module",
   "files": [
     "dist"
